refactor(chat): move ChatPage resource handling into a controller hook
The side-panel resource code in ChatPage.tsx -- panel tabs, message
search, source/issue link selection and persistence, file / folder /
artifact / diff opening, file save, screen snips, uploads, the optimizer
result hand-off and drag-and-drop -- moves into
pages/chat/useChatPageResourcesController.tsx. The page calls the hook
once where that code used to sit and destructures the values it still
reads. handleSubagentPanelOpen, sessionTitles, handleFork and
handlePlanFromHere stay in the page.

The moved text is main's bytes, with the three exceptions below. Two
mechanical edits: the page state it reads arrives as hook arguments
(composer and capture ports, showActionError), and the dependency arrays
name the refs and setters the hook now receives (referentially stable, so
nothing re-fires). currentProjectRef is declared before the hook call
because the call site reads it eagerly; its value is computed from the
same slots lookup every render as before.

First exception, because the move is what made it visible: takeScreenshot
ended in `} catch { /* user cancelled */ }`, and that comment is wrong. A
cancelled capture is not an error -- POST /api/screenshot answers it with
200 and `{"path": ""}` (dashboard/handlers/files.py), which the `if (path)`
guard directly above already absorbs. So nothing reaching that catch was
ever a cancellation: it swallowed the 400 off macOS, the 120-second
capture timeout, and any request that never reached the gateway, leaving
the user clicking Screenshot with no attachment and no notice. It now
reports through showActionError (added to the dependency array).

Second exception, answering this PR's UX review: that notice names the
action rather than only restating the transport error. A bare "screenshot
timed out" above the composer does not say which click failed, so the
message goes through a new `screenshot_failed_reason` key -- "Screenshot
failed: {{reason}}" -- on the same shape as the existing
send_failed_with_error and could_not_read_file_reason. The key is
translated in all eleven locales rather than left to fall back:
catalogParity.test.ts requires every en key in every catalog, and the
i18n gate's changed-passthrough check rejects an English value dropped
into a translated catalog, so a new key needs real translations or none.
Each value follows its own locale's conventions in the sibling key --
full-width punctuation for zh-CN and ja, the space before the colon for
fr -- and preserves {{reason}} verbatim. Two cases in
ChatPageMoreCoverage.test.tsx pin both halves, a rejected capture
surfacing and a cancelled one staying silent, and each mutation of that
pair fails one of them.

Third exception, forced by a gate rather than chosen: the composer-append
expression was `previous.trim() ? \`${previous.trimEnd()}\n\n${text}\` : text`
verbatim from main, and relocating it makes it an added line, which the
i18n gate reads at zero tolerance even though `\n\n` is structural
whitespace and not copy. It is now `[previous.trimEnd(), text].join('\n\n')`
-- the same value, in a shape the rule does not flag. Exempting the file in
eslint.i18n.config.js was the alternative and was rejected: it would blind
the copy gate to every future literal in an 826-line module.

Third slice of the ChatPage split proposed in #7255.
