+++ docs/feature-map/README.md
@@ -60,7 +60,7 @@ this area is reached from inside it unless stated otherwise.
 
 | Feature | What it is | Reach it | Page | Handler | Endpoints |
 |---|---|---|---|---|---|
-| Sessions | Multi-slot agent chat, one slot per conversation; pinned rows form a manually ordered section above automatic sorting | `/chat/:slug?` — rail **Sessions** | `pages/ChatPage.tsx`, `pages/ChatSidebar.tsx`, `pages/chat/SessionFlyout.tsx`, `pages/chat/TranscriptScrollShell.tsx` (internal split of ChatPage — the transcript scroller skeleton, no new user-facing feature), `pages/chat/ChatPageMessageContent.tsx` (internal split of ChatPage — the header menu, row-key helpers and user-bubble content renderers, no new user-facing feature), `pages/chat/useChatPageTranscriptController.tsx` (internal split of ChatPage — scroll-to-bottom, auto-follow gating, nav scrolling and the pinned-prompt banner, no new user-facing feature), `pages/chat/useChatPageSessionController.ts` (internal split of ChatPage — session identity, `?sid` URL sync, session tabs and slot auto-create, no new user-facing feature), `pages/chat/hoverHold.ts` (internal split of ChatSidebar — seat arithmetic for the hovered-row hold, no new user-facing feature), `utils/pinnedSessionOrder.ts` | `chat_handlers.py`, `ws.py` | `POST /api/chat`, `GET,POST /api/chat/slots`, `GET /api/ws` |
+| Sessions | Multi-slot agent chat, one slot per conversation; pinned rows form a manually ordered section above automatic sorting | `/chat/:slug?` — rail **Sessions** | `pages/ChatPage.tsx`, `pages/ChatSidebar.tsx`, `pages/chat/SessionFlyout.tsx`, `pages/chat/TranscriptScrollShell.tsx` (internal split of ChatPage — the transcript scroller skeleton, no new user-facing feature), `pages/chat/ChatPageMessageContent.tsx` (internal split of ChatPage — the header menu, row-key helpers and user-bubble content renderers, no new user-facing feature), `pages/chat/useChatPageTranscriptController.tsx` (internal split of ChatPage — scroll-to-bottom, auto-follow gating, nav scrolling and the pinned-prompt banner, no new user-facing feature), `pages/chat/useChatPageSessionController.ts` (internal split of ChatPage — session identity, `?sid` URL sync, session tabs and slot auto-create, no new user-facing feature), `pages/chat/useChatPageResourcesController.tsx` (internal split of ChatPage — side-panel tabs, source/issue selection, file/folder/artifact/diff opening, uploads, screen snips and drops, no new user-facing feature), `pages/chat/hoverHold.ts` (internal split of ChatSidebar — seat arithmetic for the hovered-row hold, no new user-facing feature), `utils/pinnedSessionOrder.ts` | `chat_handlers.py`, `ws.py` | `POST /api/chat`, `GET,POST /api/chat/slots`, `GET /api/ws` |
 | Remote-bound session | A local session whose turns execute on a connected peer crew and stream back over its tunnel — local sidebar row, transcript and history, remote execution | New-chat menu → **New chat on crew** → pick a connected crew | `pages/ChatPage.tsx`, `pages/ChatSidebar.tsx`, `components/RemoteCrewChip.tsx` | `chat_handlers.py`, `handlers_instances.py`, `handlers/core.py`, `remote_relay.py`, `remote_mirror.py` | `POST /api/chat/slots` (`instance_id`), `POST /api/chat?relay=1`, `GET /api/instances/{id}/capabilities`, `GET /api/version` |
 | Session folders | User-defined folders grouping session rows | Sidebar folder header → drag a row | `pages/chat/FolderPanel.tsx` | `chat_folders.py` | `GET,POST /api/chat/folders`, `PATCH /api/chat/slots/{slot}/folder` |
 | Session tags | Colored labels on sessions, filterable | Sidebar row context menu → Tags | `pages/chat/SessionFlyout.tsx` | `chat_tags.py` | `GET,POST /api/chat/tags`, `PUT /api/chat/slots/{slot}/tags` |
