+++ website/src/i18n/locales/de.json
@@ -9642,6 +9642,7 @@
       "send_failed_with_error": "Senden fehlgeschlagen: {{error}}",
       "delivery_unconfirmed": "Zustellung nicht bestätigt — dein Text steht wieder im Eingabefeld; prüfe den Verlauf, bevor du erneut sendest.",
       "could_not_read_file_reason": "{{path}} konnte nicht gelesen werden ({{reason}})",
+      "screenshot_failed_reason": "Screenshot fehlgeschlagen: {{reason}}",
       "http_status": "HTTP {{status}}",
       "could_not_rename_session": "Die Sitzung konnte nicht umbenannt werden",
       "could_not_generate_title": "Es konnte kein Titel erzeugt werden",
