+++ website/src/i18n/locales/en-XA.json
@@ -9518,6 +9518,7 @@
       "send_failed_with_error": "[Şèñð ƒàìĺèð: {{error}} ············]",
       "delivery_unconfirmed": "[Ðèĺìṽèŕý ñøţ çøñƒìŕɱèð — ýøùŕ ţèẋţ ìş ƀàçķ ìñ ţĥè çøɱþøşèŕ; çĥèçķ ţĥè ţŕàñşçŕìþţ ƀèƒøŕè şèñðìñğ ìţ àğàìñ. ································]",
       "could_not_read_file_reason": "[Çøùĺð ñøţ ŕèàð {{path}} ({{reason}}) ················]",
+      "screenshot_failed_reason": "[Şçŕèèñşĥøţ ƒàìĺèð: {{reason}} ·················]",
       "http_status": "[ĤŢŢÞ {{status}} ········]",
       "could_not_rename_session": "[Çøùĺðñ'ţ ŕèñàɱè ţĥè şèşşìøñ ···················]",
       "could_not_generate_title": "[Çøùĺðñ'ţ ğèñèŕàţè à ţìţĺè ··················]",
