+++ website/src/i18n/locales/hi.json
@@ -9642,6 +9642,7 @@
       "send_failed_with_error": "भेजना विफल: {{error}}",
       "delivery_unconfirmed": "डिलीवरी की पुष्टि नहीं हुई — तुम्हारा टेक्स्ट इनपुट बॉक्स में वापस आ गया है; दोबारा भेजने से पहले बातचीत देखो।",
       "could_not_read_file_reason": "{{path}} पढ़ा नहीं जा सका ({{reason}})",
+      "screenshot_failed_reason": "स्क्रीनशॉट विफल: {{reason}}",
       "http_status": "HTTP {{status}}",
       "could_not_rename_session": "सत्र का नाम नहीं बदल सके",
       "could_not_generate_title": "शीर्षक नहीं बना सके",
